refactor class location
